refactor(frontend): extract reusable cmc services
